The face names of a trigrid or hexagrid object

Shorthand function to get the face names of an icosahedral grid or a grid linked to a facelayer.

Usage

faces(x)

# S4 method for class 'trigrid'
faces(x)

# S4 method for class 'gridlayer'
faces(x)

Arguments

x

(trigrid, hexagrid or facelayer) The grid or facelayer object.

Value

The names of the faces of the grid as a character vector.

Examples

ball <- hexagrid(2)
faces(ball)
#>  [1] "F1"  "F2"  "F3"  "F4"  "F5"  "F6"  "F7"  "F8"  "F9"  "F10" "F11" "F12"
#> [13] "F13" "F14" "F15" "F16" "F17" "F18" "F19" "F20" "F21" "F22" "F23" "F24"
#> [25] "F25" "F26" "F27" "F28" "F29" "F30" "F31" "F32" "F33" "F34" "F35" "F36"
#> [37] "F37" "F38" "F39" "F40" "F41" "F42"