Global Triangular and Hexa-Pentagonal Grids Based on Tessellated Icosahedra

The icosa package provides tools to aggregate and analyze geographic data using grids based on tessellated icosahedra. The procedures can be set to provide a grid with a custom resolution. Both the primary triangular and their inverted penta- hexagonal grids are available for implementation. Additional functions are provided to position points (latitude-longitude data) on the grids, to allow 2D and 3D plotting, use raster and vector spatial data.

Details

Note that similar to R, the package comes with absolutely no warranty. Notes about found bugs and suggestions are more than welcome!

See also

Author

Adam T. Kocsis (adam.t.kocsis@gmail.com)

Examples

# Create a triangular grid
tri <- trigrid(c(2,2))