Export trigrid class object as Wavefront .obj file

The function will take the given trigrid class object and write it's vertex, edge and face information as a .obj file

Usage

saveOBJ(x, ...)

# S4 method for class 'trigrid'
saveOBJ(x, file, scale = TRUE)

# S4 method for class 'hexagrid'
saveOBJ(x, file, scale = TRUE)

Arguments

x

A trigrid class object.

...

Arguments of class-specific methods.

file

A character path to a file to write.

scale

A logical Should the grid vertices be scaled to unit diameter? Otherwise the values in kilometers will be exported.

Value

The function has no return value.

Details

Note that hexagrid class objects are exported in their triangulated form (subfaces). The order of faces for hexagrids is not the natural (UI) order but the internal order of subfaces.

Examples

gr <- hexagrid(spacing=4)
#> Selecting hexagrid with tessellation vector: c(17).
#> Spacing: 4.072 degrees.
# example written into temporary directroy
td <- tempdir()
td
#> [1] "/tmp/RtmphiOusF"
# actual writing
saveOBJ(gr, file=file.path(td, "hexagrid.obj"))
#> Error in .local(x, ...): object 'path' not found