Spacing of cell centers

This function will return the distance between neighboring face centers.

Usage

spacing(x, ...)

# S4 method for class 'trigrid'
spacing(x, degree = TRUE)

Arguments

x

(trigrid, hexagrid) Object.

...

Arguments of class-specific methods.

degree

(logical) Should the output be returned in degrees or in kóilometers?

Value

A named numeric vector, one value for every for every neighboring cell pair.

Details

The value for every pair is given in either degrees or kilometers depending on degree.

Examples

h <- hexagrid(3)
spacing(h)
#>   F1--F3   F1--F4   F1--F2   F1--F5   F1--F6 F10--F23 F10--F11 F10--F21 
#> 21.14498 21.14498 21.14498 21.14498 21.14498 24.48283 23.31925 24.48283 
#>  F10--F9 F12--F24 F12--F11 F12--F26 F12--F13 F14--F29 F14--F15 F14--F27 
#> 23.31925 24.48283 23.31925 24.48283 23.31925 24.48283 23.31925 24.48283 
#> F14--F13 F16--F30 F16--F15 F16--F17  F16--F7  F17--F7 F17--F32 F17--F33 
#> 23.31925 24.48283 23.31925 24.48283 23.31925 23.31925 24.48283 23.31925 
#>  F18--F7 F17--F18 F18--F33  F19--F8 F19--F20 F18--F19 F19--F35 F19--F34 
#> 23.31925 21.14498 23.31925 21.14498 21.14498 21.14498 21.14498 21.14498 
#>   F2--F3   F2--F8   F2--F9   F2--F6   F2--F7  F20--F9 F20--F21 F20--F35 
#> 24.48283 21.14498 23.31925 24.48283 23.31925 23.31925 21.14498 24.48283 
#> F20--F36  F21--F9 F21--F37 F21--F36 F10--F22 F22--F23 F21--F22 F22--F38 
#> 23.31925 23.31925 24.48283 23.31925 21.14498 21.14498 21.14498 21.14498 
#> F22--F37 F23--F11 F23--F24 F23--F39 F24--F11 F24--F40 F24--F39 F12--F25 
#> 21.14498 23.31925 21.14498 23.31925 23.31925 24.48283 23.31925 21.14498 
#> F24--F25 F25--F26 F25--F40 F25--F41 F26--F13 F26--F42 F27--F13 F26--F27 
#> 21.14498 21.14498 21.14498 21.14498 23.31925 23.31925 23.31925 21.14498 
#> F27--F42 F14--F28 F28--F29 F27--F28 F28--F43 F28--F44 F29--F15 F29--F30 
#> 23.31925 21.14498 21.14498 21.14498 21.14498 21.14498 23.31925 21.14498 
#> F29--F45   F3--F4  F10--F3  F3--F11   F3--F9 F30--F15 F30--F45 F16--F31 
#> 23.31925 24.48283 21.14498 23.31925 23.31925 23.31925 23.31925 21.14498 
#> F30--F31 F17--F31 F31--F46 F31--F32 F32--F46 F32--F47 F32--F33 F18--F34 
#> 21.14498 21.14498 21.14498 21.14498 24.48283 23.31925 23.31925 24.48283 
#> F34--F49 F34--F33 F34--F35 F34--F50 F35--F36 F35--F51 F35--F50 F37--F53 
#> 21.14498 23.31925 24.48283 23.31925 23.31925 21.14498 23.31925 23.31925 
#> F37--F52 F37--F36 F37--F38 F38--F54 F38--F53 F23--F38 F38--F39  F4--F11 
#> 21.14498 23.31925 24.48283 21.14498 23.31925 24.48283 23.31925 23.31925 
#>  F12--F4  F4--F13 F40--F39 F40--F56 F26--F41 F41--F42 F40--F41 F41--F56 
#> 21.14498 23.31925 23.31925 23.31925 24.48283 23.31925 24.48283 23.31925 
#> F43--F44 F43--F59 F27--F43 F43--F42 F44--F59 F29--F44 F44--F45 F30--F46 
#> 24.48283 23.31925 24.48283 23.31925 23.31925 24.48283 23.31925 24.48283 
#> F46--F45 F46--F47 F48--F47 F32--F48 F48--F33 F48--F49 F49--F33 F49--F65 
#> 23.31925 23.31925 23.31925 21.14498 23.31925 24.48283 23.31925 24.48283 
#> F49--F50   F5--F6  F14--F5  F5--F15   F4--F5  F5--F13 F51--F52 F51--F36 
#> 23.31925 24.48283 21.14498 23.31925 24.48283 23.31925 24.48283 23.31925 
#> F51--F66 F51--F50 F52--F68 F52--F53 F52--F36 F54--F69 F54--F53 F54--F55 
#> 24.48283 23.31925 24.48283 23.31925 23.31925 24.48283 23.31925 24.48283 
#> F54--F39 F55--F39 F40--F55 F55--F56 F57--F42 F41--F57 F57--F56 F58--F59 
#> 23.31925 23.31925 21.14498 23.31925 23.31925 21.14498 23.31925 23.31925 
#> F43--F58 F57--F58 F58--F42  F6--F15  F16--F6   F6--F7 F60--F59 F44--F60 
#> 21.14498 24.48283 23.31925 23.31925 21.14498 23.31925 23.31925 21.14498 
#> F60--F61 F60--F45 F61--F45 F46--F61 F61--F47 F62--F77 F62--F78 F61--F62 
#> 24.48283 23.31925 23.31925 21.14498 23.31925 24.48283 23.31925 24.48283 
#> F62--F47 F63--F79 F62--F63 F63--F78 F48--F63 F63--F47 F63--F64 F64--F79 
#> 23.31925 24.48283 21.14498 23.31925 24.48283 23.31925 21.14498 21.14498 
#> F48--F64 F64--F65 F49--F64 F65--F80 F65--F66 F65--F50 F66--F81 F66--F80 
#> 21.14498 21.14498 21.14498 23.31925 21.14498 23.31925 24.48283 23.31925 
#> F66--F50 F52--F67 F67--F68 F51--F67 F67--F81 F66--F67 F68--F53 F68--F82 
#> 23.31925 21.14498 21.14498 21.14498 21.14498 21.14498 23.31925 23.31925 
#> F69--F53 F68--F69 F69--F82 F54--F70 F69--F70 F55--F70 F70--F71 F70--F83 
#> 23.31925 21.14498 23.31925 21.14498 21.14498 21.14498 21.14498 21.14498 
#> F71--F83 F71--F84 F55--F71 F71--F56 F72--F84 F71--F72 F57--F72 F72--F56 
#> 24.48283 23.31925 24.48283 23.31925 23.31925 21.14498 24.48283 23.31925 
#> F73--F74 F73--F85 F58--F73 F72--F73 F57--F73 F74--F85 F74--F75 F74--F86 
#> 21.14498 21.14498 21.14498 21.14498 21.14498 24.48283 21.14498 23.31925 
#> F58--F74 F74--F59 F75--F77 F75--F86 F60--F75 F75--F59 F75--F76 F76--F77 
#> 24.48283 23.31925 24.48283 23.31925 24.48283 23.31925 21.14498 21.14498 
#> F62--F76 F60--F76 F61--F76 F77--F86 F77--F87 F77--F78 F79--F78 F79--F88 
#> 21.14498 21.14498 21.14498 23.31925 21.14498 23.31925 23.31925 21.14498 
#> F65--F79 F79--F80  F20--F8   F8--F9  F18--F8   F8--F7 F68--F81 F81--F82 
#> 24.48283 23.31925 24.48283 23.31925 24.48283 23.31925 24.48283 23.31925 
#> F81--F80 F83--F84 F69--F83 F83--F82 F85--F86 F85--F91 F72--F85 F85--F84 
#> 23.31925 23.31925 24.48283 23.31925 23.31925 21.14498 24.48283 23.31925 
#> F87--F86 F87--F78 F87--F88 F88--F78 F88--F89 F88--F80 F89--F82 F81--F89 
#> 23.31925 23.31925 24.48283 23.31925 24.48283 23.31925 23.31925 21.14498 
#> F89--F80 F90--F84 F83--F90 F89--F90 F90--F82 F87--F91 F91--F86 F90--F91 
#> 23.31925 23.31925 21.14498 24.48283 23.31925 24.48283 23.31925 24.48283 
#> F91--F84 F91--F92 F87--F92 F88--F92 F90--F92 F89--F92 
#> 23.31925 21.14498 21.14498 21.14498 21.14498 21.14498