Tessellation guide to trigrid objects

The table includes basic properties of trigrids described with specific tessellation parameters

Usage

triguide

Format

A data.frame with 120 observations and 20 variables:

total

The total tessellation of the grid, the number of points inserted between icosahedron vertices along an edge.

level1

Level 1 tessellation.

level2

Level 2 tessellation - second value of the tessellation vector.

level3

Level 3 tessellation - third value of the tessellation vector.

level4

Level 4 tessellation - four value of the tessellation vector.

faces

The number of faces in the grid.

vertices

The number of vertices in the grid.

meanEdgeLength_deg

Mean edge length in degrees.

sdEdgeLength_deg

Standard deviation of edge length in degrees.

meanEdgeLength_km

Mean edge length in kilometers.

sdEdgeLength_km

Standard devation of edge length in kilometers.

meanArea_km2

Mean face area in square-kilometers.

sdArea_km2

Standard deviation of face area in square-kilometers.

time

Time to compute grid with an Intel Xeon E-1650 prcessor.

time_sp

Time to compute grid with an Intel Xeon E-1650 prcessor, with the 'sp' member.

size

The size of the grid in bytes.

size_sp

The size of the grid object in bytes, with the 'sp' member.

timeLocate_5000

Time to locate 5000 points with an Intel Xeon E-1650 in seconds (wall time).

meanSpacing_deg

Mean spacing in degrees.

sdSpacing_deg

Standard deviation of spacing in degrees.

Details

This is version 0.2, now including the spacing of the grids.